Refactor: Moved classes
