+++ RoundGame/Deck.cs
@@ -1,0 +1,10 @@
+﻿namespace RoundGame
+{
+    public class Deck
+    {
+        public virtual Card GetRandomCard()
+        {
+            return null;
+        }
+    }
+}
